LayoutTests/fast/page-color-sampling/color-sampling-ignores-large-sticky-container.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=50 obscuredInset.left=50 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            margin: 0;
            font-family: system-ui;
        }

        .content {
            position: sticky;
            width: 100%;
            height: 2000px;
            background: #eee;
        }

        h1 {
            width: 100%;
            text-align: center;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(50, 50, 50, 50);
        await UIHelper.ensurePresentationUpdate();

        colors = await UIHelper.fixedContainerEdgeColors();
        shouldBeNull("colors.top");
        shouldBeEqualToString("colors.left", "rgb(238, 238, 238)");
        shouldBeEqualToString("colors.right", "rgb(238, 238, 238)");
        shouldBeNull("colors.bottom");

        finishJSTest();
    });
    </script>
</head>
<body>
    <div class="content">
        <h1>Hello world</h1>
    </div>
</body>
</html>
